requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==1.5.3
matplotlib==3.7.1
mysql-connector-python==8.0.32
langchain==0.0.215
openai==0.27.0
chromadb==0.3.20
python-dotenv==0.21.1
